Add updated outbound generaltab to use consistent pattern for content area
